LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/tab-size-expected.txt


PASS Can set 'tab-size' to CSS-wide keywords
FAIL Can set 'tab-size' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'tab-size' to a number assert_equals: expected 2 but got 1
FAIL Can set 'tab-size' to a length assert_equals: expected "px" but got "em"
PASS Setting 'tab-size' to a percent throws TypeError
PASS Setting 'tab-size' to a time throws TypeError
PASS Setting 'tab-size' to an angle throws TypeError
PASS Setting 'tab-size' to a flexible length throws TypeError
PASS Setting 'tab-size' to a URL throws TypeError
PASS Setting 'tab-size' to a transform throws TypeError
